% Generated by roxygen2 (4.0.1): do not edit by hand
\encoding{utf-8}
\name{autoPCA}
\alias{autoPCA}
\title{autoPCA}
\usage{
autoPCA(env.stack, nfactors = 0)
}
\arguments{
\item{env.stack}{a \code{rasterStack} object that contain the environment variable}

\item{nfactors}{Number of factors to extract, if it's 0 that method will auto descion the number. Default is 0}
}
\value{
\code{rasterStack} object
}
\description{
easily used PCA analysis
}
\details{
This method implemented an easily used PCA analysis method
}

