% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rescale}
\alias{rescale}
\title{rescale}
\usage{
rescale(raster.object)
}
\arguments{
\item{raster.object}{an object of RasterStack or RasterLayer class}
}
\value{
an object of RasterStack or RasterLayer that rescaled.
}
\description{
rescale the RasterStack or RasterLayer values to min:0 max:1
}

