% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rescaleLayer}
\alias{rescaleLayer}
\title{rescaleLayer}
\usage{
rescaleLayer(raster.layer)
}
\arguments{
\item{raster.layer}{an object of RasterLayer class}
}
\value{
an object of RasterLayer that rescaled.
}
\description{
rescale the RasterLayer values to min:0 max:1
}

