\name{sdmvspecies-package}

\alias{sdmvspecies-package}

\alias{sdmvspecies}

\docType{package}

\title{
create virtual species for species distribution modelling
}

\description{
this package include methods to create virtual species distribution map.

those map can used for species distribution model or ecological niche model study.
}

\details{
\tabular{ll}{
Package: \tab sdmvspecies\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2014-01-05\cr
License: \tab GPL\cr
}
With input of raster variables, this package output the distribution map of virutal species.

You can use serive methods to create virtual species, all depends on you.
}

\author{
Xiaoquan Kong, Renyuan Duan, Minyi Huang

Maintainer: Howl Anderson <u1mail2me@gmail.com>
}

\references{
This package depends on raster package, if you familiar with that, all things will be very easy.
}

\keyword{virtual species}
\keyword{species distribution modelling}
