% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\encoding{utf-8}
\name{sdmvspecies}
\alias{sdmvspecies}
\alias{sdmvspecies-package}
\title{SDMvspcies}
\description{
Species Distribution Modelling (SDM) tools for Virtual Species (vspecies)
}
\details{
This package contain some useful tools for create and study virtual species in SDM

SDMvspcices is a tools package for creating virtual species in Species Distribution Modelling (SDM)
It contains servel algorithms (methods) that already report and used in current vritual species study or application.
Also many useful tools are include to help user development new algorithms (methods) and study virtual species.
}
\author{
Xiaoquan Kong

Renyan Duan

Minyi Huang
}

