% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_PCA.R
\encoding{utf-8}
\name{autoPCA}
\alias{autoPCA}
\title{autoPCA}
\usage{
autoPCA(env.stack, nfactors)
}
\arguments{
\item{env.stack}{a \code{rasterStack} object that contain the environment variable}

\item{nfactors}{Number of factors to extract}
}
\value{
\code{rasterStack} object
}
\description{
easily used PCA analysis
}
\details{
This method implemented an easily used PCA analysis method
}

