% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale.R
\name{rescaleStack}
\alias{rescaleStack}
\title{rescaleStack}
\usage{
rescaleStack(raster.stack)
}
\arguments{
\item{raster.stack}{an object of RasterStack class}
}
\value{
an object of RasterStack that rescaled.
}
\description{
rescale the RasterStack values to min:0 max:1
}

